requirements/base.txt
asgiref==3.3.4
Django==3.2.2
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
django-environ==0.4.5
dependency-injector==4.32.2
psycopg2
django-cors-headers